Updated my part of the app in an effort to increase performance for the next checkpoint. Previously, my part made a browser API call for whatever the heck people put in the text field, and then would parse the JSON that the call returns for the placeIDs. After that, it would take those placeIDs and make a Places Android API call for each one to get the Place object that they referred to.
Since we needed to "improve performance" for this checkpoint, my part was a pretty easy candidate since it made all those calls.

I made it so that my part now only does the initial browser API call, now additionally parsing for not just the placeID, but also the Lat, Long, and Name of the places. I added some ArrayLists to the FoodMap fragment to accomodate these, so now the FoodMapFragment basically has two options when Map it! is pressed. Either the mPlaces arrayList will have stuff in it, or the other three will have stuff in them. If mPlaces has stuff, then it will act like it did before, but if not, then it will use those three arrayLists to make the markers and save stuff instead.

I also added a small check to make sure that the restaurant ID in one of the DbOperator's methods wasn't null, since because of testing I now have a null pointer in my database, so the app was crashing every time it tried to save something. Should be all good now though.